Handle empty string for simple types
- Add and register a new formatter for the Decimal type.
- Register serializer and deserializer middleware for Decimals:
  - Convert empty string to `undefined` if nullable.
  - Convert empty string to `0` if non-nullable.
- Apply similar handling for other simple types:
  - Convert empty string to `undefined` if nullable.
- Add tests covering serialization of Decimals and other simple types
receiving an empty string.

+++ typescript-react/src/components/FieldRegistry/FieldRegistryContext.ts
@@ -12,7 +12,7 @@ export interface IFields {
   Decimal: React.ComponentType<IExternalFormField<any, any, DecimalStr>>;
   Double: React.ComponentType<IExternalFormField<any, any, Double>>;
   EnumSelect: React.ComponentType<IExternalFormField<any, any, any>>;
-  Float: React.ComponentType<IExternalFormField<any, any, Double>>;
+  Float: React.ComponentType<IExternalFormField<any, any, Float>>;
   Integer: React.ComponentType<IExternalFormField<any, any, Int>>;
   Link: React.ComponentType<IExternalFormField<any, any, string>>;
   Long: React.ComponentType<IExternalFormField<any, any, Long>>;
